refactor(next/web): rename TicketContext to mixed-ticket
